Perform movement calculations using floats
Using floats made dividing the speed by a factor much easier to
calculate.

Also, as I've used ptrkeys as my primary pointing mechanism the
importance of being able to use it single-handedly has been made obvious
to me, and the changes to the bindings reflect this.

+++ pk.h
@@ -4,10 +4,10 @@
 #include <X11/Xlib.h>
 
 typedef struct {
-	unsigned int basespeed;
+	double basespeed;
 	unsigned int dir;  // Bits from enum Direction.
-	unsigned int mul;
-	int xrem, yrem; // Subunit remainders.
+	double mul;
+	double xrem, yrem; // Subunit remainders.
 	int xcont, ycont; // Continuing a movement?
 } Movement;
 
@@ -15,7 +15,7 @@ typedef union {
 	int i;
 	unsigned int ui;
 	unsigned long ul;
-	float f;
+	double f;
 	const void *v;
 } Arg;
